+++ filefinder/tests/test_cmip.py
@@ -16,8 +16,8 @@ def test_parse_ens_cmip5():
             ("file4", "r1i2p1"),
             ("file5", "r1i1p2"),
         ],
-        columns=("filename", "ens"),
-    )
+        columns=("path", "ens"),
+    ).set_index("path")
 
     fc = FileContainer(df)
 
@@ -29,8 +29,8 @@ def test_parse_ens_cmip5():
             ("file4", "r1i2p1", "1", "2", "1"),
             ("file5", "r1i1p2", "1", "1", "2"),
         ],
-        columns=("filename", "ens", "r", "i", "p"),
-    )
+        columns=("path", "ens", "r", "i", "p"),
+    ).set_index("path")
 
     result = parse_ens(fc)
     assert isinstance(result, FileContainer)
@@ -50,8 +50,8 @@ def test_parse_ens_cmip6():
             ("file4", "r1i2p1f1"),
             ("file5", "r1i1p2f1"),
         ],
-        columns=("filename", "ens"),
-    )
+        columns=("path", "ens"),
+    ).set_index("path")
 
     fc = FileContainer(df)
 
@@ -63,8 +63,8 @@ def test_parse_ens_cmip6():
             ("file4", "r1i2p1f1", "1", "2", "1", "1"),
             ("file5", "r1i1p2f1", "1", "1", "2", "1"),
         ],
-        columns=("filename", "ens", "r", "i", "p", "f"),
-    )
+        columns=("path", "ens", "r", "i", "p", "f"),
+    ).set_index("path")
 
     result = parse_ens(fc)
     assert isinstance(result, FileContainer)
@@ -73,7 +73,7 @@ def test_parse_ens_cmip6():
 
 
 def test_create_ensnumber():
-    columns = ("filename", "model", "exp", "table", "varn", "ens")
+    columns = ("path", "model", "exp", "table", "varn", "ens")
 
     common = ("exp", "table", "varn")
 
@@ -88,12 +88,12 @@ def test_create_ensnumber():
     df = pd.DataFrame.from_records(
         records,
         columns=columns,
-    )
+    ).set_index("path")
 
     expected_df = pd.DataFrame.from_records(
         records,
         columns=columns,
-    )
+    ).set_index("path")
     expected_df["ensnumber"] = (0, 1, 2, 0, 1)
 
     fc = FileContainer(df)
@@ -106,7 +106,7 @@ def test_create_ensnumber():
 
 def test_ensure_unique_grid():
 
-    columns = ("model", "exp", "table", "varn", "ens", "grid")
+    columns = ("path", "model", "exp", "table", "varn", "ens", "grid")
 
     # VALID_GRIDS = ("gn", "gr", "gr1", "gm")
 
@@ -114,22 +114,22 @@ def test_ensure_unique_grid():
 
     df = pd.DataFrame.from_records(
         [
-            ("CESM2", *common, "gr"),
-            ("CESM2", *common, "gn"),
-            ("CESM2", *common, "gm"),
-            ("UKESM", *common, "gr"),
-            ("UKESM", *common, "gr1"),
+            ("CESM2_gr", "CESM2", *common, "gr"),
+            ("CESM2_gn", "CESM2", *common, "gn"),
+            ("CESM2_gm", "CESM2", *common, "gm"),
+            ("UKESM_gr", "UKESM", *common, "gr"),
+            ("UKESM_gr1", "UKESM", *common, "gr1"),
         ],
         columns=columns,
-    )
+    ).set_index("path")
 
     expected = pd.DataFrame.from_records(
         [
-            ("CESM2", *common, "gn"),
-            ("UKESM", *common, "gr"),
+            ("CESM2_gn", "CESM2", *common, "gn"),
+            ("UKESM_gr", "UKESM", *common, "gr"),
         ],
         columns=columns,
-    )
+    ).set_index("path")
 
     result = ensure_unique_grid(df)