+++ CHANGELOG.md
@@ -11,6 +11,12 @@
   longer be passed by position ([#99](https://github.com/mathause/filefinder/pull/99)).
 - `FileFinder` now raises an error if an invalid `"{placeholder}"` is used
    ([#99](https://github.com/mathause/filefinder/pull/99)).
+
+- Changes to `FileContainer`:
+
+  - Renamed the `"filename"` column to `"path"` and made it a `pd.Index`, thus removing
+    this column from the underlying `DataFrame` ([#113](https://github.com/mathause/filefinder/pull/113)).
+
 - Explicitly test on python 3.13 ([#103](https://github.com/mathause/filefinder/pull/103)).
 - Drop support for python 3.9 ([#102](https://github.com/mathause/filefinder/pull/102)).